+++ src/components/SignInForm/SubmitButton.tsx
@@ -1,10 +1,16 @@
 import { FC } from 'react';
 import { Button } from '../Button';
 
-export const SubmitButton: FC<{ title: string; onSubmit: () => void }> = (
+export const SubmitButton: FC<{ title: string; onSubmit?: () => void }> = (
   props
 ) => (
-  <Button type="submit" fullWidth variant="contained" onClick={props.onSubmit}>
+  <Button
+    data-testid="submit-btn"
+    type="submit"
+    fullWidth
+    variant="contained"
+    onClick={props.onSubmit}
+  >
     {props.title}
   </Button>
 );