+++ src/components/SignInForm/index.ts
@@ -1,6 +1,5 @@
-export { SignInCard } from './SignInCard';
-export { SignInContainer } from './SignInContainer';
+import SignInForm from './SignInForm';
 
-export { SignInFormBox } from './SignInFormBox';
-export { SubmitButton } from './SubmitButton';
-export { SignUpLink } from './SignUpLink';
+export type { SignInFormInput } from './types';
+
+export { SignInForm };